final batch of RSession monolith split
