+++ src/cpp/r/include/r/session/RSession.hpp
@@ -32,7 +32,7 @@
 
 namespace rstudio {
 namespace core {
-	class Error;
+	class Error ;
    class Settings;
 } 
 }
