+++ src/cpp/r/include/r/session/RSessionUtils.hpp
@@ -60,7 +60,11 @@ core::FilePath rSourcePath();
 
 core::FilePath rHistoryDir();
 
+core::FilePath rEnvironmentDir();
+
 core::FilePath sessionScratchPath();
+
+core::FilePath scopedScratchPath();
 
 core::FilePath clientStatePath();
 
@@ -68,11 +72,17 @@ core::FilePath projectClientStatePath();
 
 core::FilePath startupEnvironmentFilePath();
 
+core::FilePath suspendedSessionPath();
+
 std::string sessionPort();
 
 std::string rCRANRepos();
 
 bool restoreWorkspace();
+
+bool useInternet2();
+
+bool alwaysSaveHistory();
 
 // suppress output in scope
 class SuppressOutputInScope
