+++ src/cpp/r/session/RSuspend.hpp
@@ -32,6 +32,9 @@ bool suspend(const RSuspendOptions& options,
              const core::FilePath& suspendedSessionPath,
              bool disableSaveCompression,
              bool force);
+bool suspended();
+
+void saveClientState(ClientStateCommitType commitType);
 
 class SerializationCallbackScope : boost::noncopyable
 {
